Tests now check test commands for validity

+++ Tests/test_LaunchedApp.cpp
@@ -1,5 +1,6 @@
 #include "JuceHeader.h"
 #include "LaunchedApp.h"
+#include "AppLauncher.h"
 #include "WindowFocus.h"
 #include "XWindowInterface.h"
 
@@ -11,27 +12,46 @@ public:
     void runTest() override
     {
         using namespace juce;
-        beginTest("echo test");
+        String output;
+
+        if(AppLauncher::testCommand("echo"))
+        {
+            beginTest("echo test");
+            LaunchedApp echo("echo (test)");
+            echo.waitForProcessToFinish(1000);
+            expect(!echo.isRunning(), 
+                "Echo process still running, but should be finished.");
+            String expected("(test)");
+            String output = echo.getProcessOutput().trim();
+            expectEquals(output, expected,
+                    "Echo process output was incorrect.");
+            expect(!echo.kill(),
+                    "Killing finished process didn't fail");
+        }
+        else
+        {
+            logMessage("echo is not a valid command, skipping test");
+        }
+
+        if(AppLauncher::testCommand("top"))
+        {
+            beginTest("top test");
+            LaunchedApp top("top");
+            system("sleep 1");
+            expect(top.isRunning(), "\"top\" process is not running.");
+            output = top.getProcessOutput();
+            expect(output.isEmpty(), String("Unexpected process output ") 
+                    + output);
+            expect(top.kill(), "Failed to kill process.");
+        }
+        else
+        {
+            logMessage("top is not a valid command, skipping test");
+        }
         
-        LaunchedApp echo("echo (test)");
-        echo.waitForProcessToFinish(1000);
-        expect(!echo.isRunning(), 
-        	"Echo process still running, but should be finished.");
-        String expected("(test)");
-        String output = echo.getProcessOutput().trim();
-        expectEquals(output, expected,
-                "Echo process output was incorrect.");
-        expect(!echo.kill(),
-		"Killing process succeeded when process should have been already dead");
-        
-        beginTest("top test");
-        LaunchedApp top("top");
-        system("sleep 1");
-        expect(top.isRunning(), "\"top\" process is not running.");
-        output = top.getProcessOutput();
-        expect(output.isEmpty(), String("Unexpected process output ") + output);
-        expect(top.kill(), "Failed to kill process.");
-        
+        String badCommand("DefinitelyNotAValidLaunchCommand");
+        expect(!AppLauncher::testCommand(badCommand),
+                "DefinitelyNotAValidLaunchCommand should have been invalid.");
         beginTest("bad command handling");
         LaunchedApp bad("DefinitelyNotAValidLaunchCommand");
         expect(!bad.isRunning(),
@@ -41,20 +61,27 @@ public:
         expectEquals(String(bad.getExitCode()), String("0"),
 			"Bad process error code should have been 0.");
         
-        beginTest("window activation");
-        LaunchedApp winApp("xclock");
-        system("sleep 1");
-        expect(winApp.isRunning(),
-		"Launched terminal process not running.");
-        winApp.activateWindow();
-	MessageManager::getInstance()->runDispatchLoopUntil(5000);
-        expect(!WindowFocus::isFocused(),
-		"pocket-home window should not be focused.");
-        winApp.kill();
-        XWindowInterface xwin;
-        xwin.activateWindow(xwin.getPocketHomeWindow());
-        expect(!winApp.isRunning(),
-			"xclock process should be dead.");
+        if(AppLauncher::testCommand("xclock"))
+        {
+            beginTest("window activation");
+            LaunchedApp winApp("xclock");
+            system("sleep 1");
+            expect(winApp.isRunning(),
+            "Launched terminal process not running.");
+            winApp.activateWindow();
+            MessageManager::getInstance()->runDispatchLoopUntil(5000);
+            expect(!WindowFocus::isFocused(),
+            "pocket-home window should not be focused.");
+            winApp.kill();
+            XWindowInterface xwin;
+            xwin.activateWindow(xwin.getPocketHomeWindow());
+            expect(!winApp.isRunning(),
+                "xclock process should be dead.");
+        }
+        else
+        {
+            logMessage("xclock is not a valid command, skipping test");
+        }
     }
 };
